Cita->clear methods to mapeo data,all models id to long

+++ src/main/java/com/timora/app/models/Configuracion.java
@@ -17,7 +17,7 @@ public class Configuracion {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_configuracion")
-    private Integer idConfiguracion;
+    private Long idConfiguracion;
 
     // Relación 1 a 1 con Usuario
     @OneToOne
